@@ -16,12 +16,13 @@ repositories {
 
 
 android {
-    compileSdkVersion 24
+    compileSdkVersion 28
     buildToolsVersion "24.0.2"
     flavorDimensions "default"
 
     defaultConfig {
         applicationId "com.AndroTools.FreeProVPN"
+        multiDexEnabled true
         minSdkVersion 17
         targetSdkVersion 24
         versionCode 50
@@ -69,21 +70,30 @@ android {
 
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:24.2.1'
-    compile 'com.android.support:support-v4:24.2.1'
-    compile 'com.android.support:design:24.2.1'
-    compile 'com.android.support:recyclerview-v7:24.2.1'
-    compile 'com.android.support:cardview-v7:24.2.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
 
-    compile 'com.daimajia.numberprogressbar:library:1.2@aar'
-    compile 'com.amitshekhar.android:android-networking:0.2.0'
-    compile 'com.google.code.gson:gson:2.7'
-    compile 'org.mapsforge:mapsforge-map-android:0.6.1'
-    compile 'com.caverock:androidsvg:1.2.2-beta-1'
+    implementation 'com.android.support:multidex:1.0.3'
+
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'com.google.code.gson:gson:2.7'
+    implementation 'com.google.android.gms:play-services-ads:19.4.0'
+    implementation 'com.google.android.gms:play-services-analytics:17.0.0'
+
+    implementation 'com.daimajia.numberprogressbar:library:1.2@aar'
+    implementation 'com.amitshekhar.android:android-networking:0.2.0'
+
+    implementation 'org.mapsforge:mapsforge-map-android:0.6.1'
+    implementation 'com.caverock:androidsvg:1.2.2-beta-1'
+
     testCompile 'junit:junit:4.12'
-    compile('com.crashlytics.sdk.android:crashlytics:2.6.5@aar') {
+    compile 'junit:junit:4.12'
+
+    //testImplementation 'junit:junit:4.12'
+    implementation('com.crashlytics.sdk.android:crashlytics:2.6.5@aar') {
         transitive = true
     }
-    compile 'com.google.android.gms:play-services-analytics:10.0.1'
 }
