@@ -4,7 +4,8 @@ import android.app.Activity;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
+
+import androidx.appcompat.app.AlertDialog;
 
 import com.AndroTools.FreeProVPN.R;
 import com.AndroTools.FreeProVPN.util.NetworkState;
