@@ -6,7 +6,8 @@
 package de.blinkt.openvpn.core;
 
 import android.os.Build;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.AndroTools.FreeProVPN.BuildConfig;
 
