@@ -7,19 +7,35 @@ import android.os.Bundle;
 import android.widget.TextView;
 
 import com.AndroTools.FreeProVPN.R;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.initialization.InitializationStatus;
+import com.google.android.gms.ads.initialization.OnInitializationCompleteListener;
 
 public class AboutActivity extends BaseActivity {
+    private AdView mAdView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_about);
+
+        MobileAds.initialize(this, new OnInitializationCompleteListener() {
+            @Override
+            public void onInitializationComplete(InitializationStatus initializationStatus) {
+            }
+        });
+        mAdView = findViewById(R.id.adView);
+        AdRequest adRequest = new AdRequest.Builder().build();
+        mAdView.loadAd(adRequest);
+
         try {
             PackageInfo pinfo = getPackageManager().getPackageInfo(getPackageName(), 0);
             int versionNumber = pinfo.versionCode;
             String versionName = pinfo.versionName;
-            String appName =  AboutActivity.this.getString(R.string.app_name);
-            TextView versionText = (TextView)findViewById(R.id.appVersion);
+            String appName = AboutActivity.this.getString(R.string.app_name);
+            TextView versionText = (TextView) findViewById(R.id.appVersion);
             versionText.setText(
                     String.format("%s version %s build %d",
                             appName,
