@@ -1,9 +1,10 @@
 package com.AndroTools.FreeProVPN.activity;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.DisplayMetrics;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.AndroTools.FreeProVPN.R;
 import com.AndroTools.FreeProVPN.database.DBHelper;
