@@ -2,13 +2,14 @@ package com.AndroTools.FreeProVPN.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.AndroTools.FreeProVPN.R;
 import com.AndroTools.FreeProVPN.activity.ServerActivity;
