@@ -1,13 +1,14 @@
 package com.AndroTools.FreeProVPN.adapter;
 
 import android.content.Context;
-import android.support.v4.content.ContextCompat;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.core.content.ContextCompat;
 
 import com.AndroTools.FreeProVPN.R;
 import com.AndroTools.FreeProVPN.activity.BaseActivity;
